services/agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-core>=0.3.37",
    "langchain-ollama>=0.2.3",
    "langchain-openai>=0.3.6",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.2.74",
    "langgraph-cli[inmem]>=0.1.73",
    "loguru>=0.7.3",
    "openai>=1.64.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.2",
    "requests>=2.32.3",
    "sentence-transformers>=3.4.1",
    "streamlit>=1.42.2",
    "tabulate>=0.9.0",
    "toolbox-langchain>=0.1.0",
    "watchdog>=6.0.0",
]
